✅ test: split monolithic work.test.ts into per-type test files
Mirror the WorkModel per-type source split in its tests: move the single
1697-line work.test.ts into work/__tests__/ as task / document / linear /
github / queries test files, sharing setup + snapshot/summary helpers via
_fixtures.ts. Each file gets its own isolated PGlite instance, so the fixed
seed ids never collide. No test bodies changed — all 26 cases still pass.
